requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-ipware==2.1.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
asgiref==3.2.3
Click==7.0
dataclasses==0.7
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-ipware==2.1.0
djangorestframework==3.11.0
fastapi==0.47.1
h11==0.9.0
httptools==0.0.13
pkg-resources==0.0.0
pydantic==1.4
pytz==2019.3
sqlparse==0.3.0
starlette==0.12.9
uvicorn==0.11.2
uvloop==0.14.0
websockets==8.1
asgiref==3.2.3
Click==7.0
dataclasses==0.7
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-ipware==2.1.0
djangorestframework==3.11.0
fastapi==0.47.1
h11==0.9.0
httptools==0.0.13
pkg-resources==0.0.0
pydantic==1.4
pytz==2019.3
SQLAlchemy==1.3.13
sqlparse==0.3.0
starlette==0.12.9
uvicorn==0.11.2
uvloop==0.14.0
websockets==8.1
